+++ core/detections/dao/index.js
@@ -275,8 +275,11 @@ async function timeAggregatedQuery (start, end, streams, streamsOnlyPublic, clas
  */
 async function queryDetectionsSummary (filters, options) {
   const opts = await defaultQueryOptions(filters, options)
-  opts.include.forEach(i => {
-    i.attributes = []
+  const includesWithoutAdditionalColumns = opts.include.map(i => {
+    return {
+      ...i,
+      attributes: []
+    }
   })
 
   /**
@@ -288,7 +291,7 @@ async function queryDetectionsSummary (filters, options) {
       [sequelize.literal('COUNT(1)::integer'), 'count']
     ],
     where: opts.where,
-    include: opts.include,
+    include: includesWithoutAdditionalColumns,
     group: ['"Detection"."review_status"'],
     raw: true
   })